fix: casting double to integer at line 72 in readDouble method

+++ src/util/CustomScanner.java
@@ -69,7 +69,7 @@ public class CustomScanner {
                 continue;
             }
 
-            option = (Integer) response.obj;
+            option = (Double) response.obj;
 
             break;
         }
